requirements.txt
streamlit
langdetect
requests
python-dotenv
google-generativeai
google-genai
pydantic
dropbox
xmltodict
